requirements.txt
bidict
lightning-bolts
pandas
ipython
ipykernel
jupyter
yellowbrick
nvidia-ml-py3  # (import pynvml) track GPU usage within python
pylint
black
flake8
torchviz
transformers
tokenizers
torchview
torchdata